refactor: delete shadow chatty_commander.tools package
The shadow tools/ package caused the original ..tools.X wrong-package
import bug (69a124d0). Contents relocated: api-docs cluster ->
cli/api_docs/, validate_config + run_tests_with_coverage -> cli/,
bridge_nodejs -> advisors/tools/, browser_analyst merged into the
canonical advisors/tools module. All references updated (Makefile,
tests); zero chatty_commander.tools imports remain.
