+++ src/chatty_commander/advisors/tools/bridge_nodejs.py
