+++ src/chatty_commander/cli/api_docs/builder.py
