+++ src/chatty_commander/cli/api_docs/cli.py
