+++ src/chatty_commander/cli/api_docs/generate_api_docs.py
